refactor: move `Node::load`/`Node::save` into separate file.
We will be introducing a new version of `Node` to accommodate varying
node sizes. `v1.rs` encapsulates the logic that's specific to V1, and
`v2.rs` will be introduced in an upcoming commit containing V2-specific
logic.
